Week09/Day2/DailyChallenge/requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-groq
tavily-python
faiss-cpu
wikipedia
transformers
accelerate
sentencepiece
